requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
pycodestyle==2.9.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.12
